Development

+++ IRSensor.c
@@ -43,7 +43,7 @@ int main(int argc, char const *argv[])
     
     
     t = clock() - t; 
-    double time_taken = ((double)t)/CLOCKS_PER_SEC; // in seconds 
+    double time_taken = ((double)t)/((double)CLOCKS_PER_SEC); // in seconds 
 
     printf("The time taken for the function is %f\nThe PI value is %f\nThe value of n is %lu\nThe value of N is %d\n",time_taken,PI,n,N);
